+++ tests/testthat/test-checkValidity.R
@@ -24,46 +24,35 @@ test_that("validity checks work", {
     expect_identical(getReadLevelAssayNames(rme_withreads), "mod_prob")
     expect_identical(getReadLevelAssayNames(rme_withoutreads), character(0))
 
+    ## Test getReadNamesBySample
+    expect_identical(getReadNamesBySample(rme_withreads),
+                     structure(
+                         list(s1_5mC = colnames(assay(rme_withreads, "mod_prob")[[1]]),
+                              s2_5mC = colnames(assay(rme_withreads, "mod_prob")[[2]]),
+                              s1_6mA = colnames(assay(rme_withreads, "mod_prob")[[3]])),
+                         source = "assay mod_prob"))
+    expect_equal(getReadNamesBySample(rme_withreads),
+                 getReadNamesBySample(rme_withoutreads),
+                 ignore_attr = TRUE)
+    expect_identical(lengths(getReadNamesBySample(rme_withreads)),
+                     c(s1_5mC = 10L, s2_5mC = 10L, s1_6mA = 10L))
+    tmp <- rme_withoutreads
+    tmp$QC <- tmp$readInfo <- NULL
+    expect_error(getReadNamesBySample(tmp),
+                 "does not contain any read-level assays or colData columns")
+
     ## Test checkSEValidity
     expect_no_error(checkSEValidity(rme_withreads))
     expect_no_error(checkSEValidity(rme_withoutreads))
 
-    expect_message(expect_message(expect_message(
-        expect_message(expect_message(expect_message(
-            expect_message(expect_message(expect_message(
-                expect_message(expect_message(expect_message(
-                    expect_message(expect_message(expect_message(
-                        checkSEValidity(rme_withreads, verbose = TRUE),
-                        "Checking assay names"), "Checking assay names")),
-                    "Checking row names"), "Checking row names")),
-                "Checking consistency of sample names"), "Checking consistency of sample names")),
-            "Read-level assay found"), "Read-level assay found")),
-        "Read-level column data found, checking consistency"), "Read-level column data found, checking consistency"))
-    expect_message(expect_message(expect_message(
-        expect_message(expect_message(expect_message(
-            expect_message(expect_message(expect_message(
-                checkSEValidity(rme_withoutreads, verbose = TRUE),
-                "Checking assay names"), "Checking assay names")),
-            "Checking row names"), "Checking row names")),
-        "Checking consistency of sample names"), "Checking consistency of sample names"))
+    suppressMessages(expect_message(checkSEValidity(rme_withreads, verbose = TRUE)))
+    suppressMessages(expect_message(checkSEValidity(rme_withoutreads, verbose = TRUE)))
 
     rme1 <- rme_withreads
     SummarizedExperiment::assay(rme1, "test") <- SummarizedExperiment::assay(rme1, "mod_prob")
     metadata(rme1)$readLevelData$assayNames <- c(metadata(rme1)$readLevelData$assayNames,
                                                  "test")
-    expect_message(expect_message(expect_message(
-        expect_message(expect_message(expect_message(
-            expect_message(expect_message(expect_message(
-                expect_message(expect_message(expect_message(
-                    expect_message(expect_message(expect_message(
-                        expect_message(expect_message(expect_message(
-                            checkSEValidity(rme1, verbose = TRUE),
-                            "Checking assay names"), "Checking assay names")),
-                        "Checking row names"), "Checking row names")),
-                    "Checking consistency of sample names"), "Checking consistency of sample names")),
-                "Read-level assay found"), "Read-level assay found")),
-            "Comparing mod_prob and test"), "Comparing mod_prob and test")),
-        "Read-level column data found, checking consistency"), "Read-level column data found, checking consistency"))
+    suppressMessages(expect_message(checkSEValidity(rme1, verbose = TRUE)))
 
     rme1 <- rme_withreads
     rownames(rme1) <- as.character(rowRanges(rme1))
@@ -119,7 +108,7 @@ test_that("validity checks work", {
     SummarizedExperiment::assay(rme1, "mod_prob")[[1]] <-
         SummarizedExperiment::assay(rme1, "mod_prob")[[1]][, 1:5]
     expect_error(checkSEValidity(rme1),
-                 "Mismatching reads for assay mod_prob and colData column QC, sample s1_5mC")
+                 "Mismatching reads for assay mod_prob and colData column readInfo, sample s1_5mC")
 
     rme1 <- rme_withreads
     SummarizedExperiment::assay(rme1, "test") <- SummarizedExperiment::assay(rme1, "mod_prob")
@@ -128,7 +117,7 @@ test_that("validity checks work", {
     SummarizedExperiment::assay(rme1, "mod_prob")[[1]] <-
         SummarizedExperiment::assay(rme1, "mod_prob")[[1]][, 1:5]
     expect_error(checkSEValidity(rme1),
-                 "Mismatching reads for assays mod_prob and test, sample s1_5mC")
+                 "Mismatching reads for assay mod_prob and test, sample s1_5mC")
 
     rme1 <- rme_withreads
     SummarizedExperiment::assay(rme1, "test") <- SummarizedExperiment::assay(rme1, "mod_prob")
@@ -139,5 +128,5 @@ test_that("validity checks work", {
     SummarizedExperiment::assay(rme1, "mod_prob")[[1]] <-
         SummarizedExperiment::assay(rme1, "mod_prob")[[1]][, sample.int(N, N)]
     expect_error(checkSEValidity(rme1),
-                 "Mismatching reads for assays mod_prob and test, sample s1_5mC")
+                 "Mismatching reads for assay mod_prob and test, sample s1_5mC")
 })